Corrected admin app frontend, logging and properties file with angular frontend

+++ BackendSpring/backend/src/main/java/org/unibl/etf/ip/backend/service/JournalService.java
@@ -3,31 +3,27 @@ package org.unibl.etf.ip.backend.service;
 import com.itextpdf.text.Document;
 import com.itextpdf.text.Paragraph;
 import com.itextpdf.text.pdf.PdfWriter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.unibl.etf.ip.backend.exceptions.MethodNotAllowedException;
+import org.unibl.etf.ip.backend.exceptions.NotFoundException;
 import org.unibl.etf.ip.backend.model.DnevnikUnosEntity;
 import org.unibl.etf.ip.backend.repository.JournalEntryRepository;
 
 import java.io.ByteArrayOutputStream;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class JournalService {
 
-    /*@Autowired
-    private JournalRepository journalRepository;*/
+    private Logger logger = LoggerFactory.getLogger(JournalService.class);
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
-
-    /*public DnevnikEntity getJournal(Integer id) throws NotFoundException {
-        return journalRepository.findById(id).orElseThrow(() -> new RuntimeException("Exception"));
-    }
-
-    public DnevnikEntity createJournal(DnevnikEntity journal) {
-       return journalRepository.save(journal);
-    }*/
 
     public List<DnevnikUnosEntity> getJournalEntries(Integer id) {
         List<DnevnikUnosEntity> entries = journalEntryRepository.findAll();
@@ -41,10 +37,21 @@ public class JournalService {
     }
 
     public DnevnikUnosEntity createJournalEntry(DnevnikUnosEntity journalEntry) {
+        logger.info("User with id " + journalEntry.getDnevnikKorisnikId() + " created new journal entry");
         return journalEntryRepository.save(journalEntry);
     }
 
-    public void deleteJournalEntry(Integer id) {
+    public void deleteJournalEntry(Integer id, Integer userId) throws NotFoundException, MethodNotAllowedException {
+        Optional<DnevnikUnosEntity> journalEntryOptional = journalEntryRepository.findById(id);
+
+        if(!journalEntryOptional.isPresent()) {
+            throw new NotFoundException();
+        }
+        DnevnikUnosEntity journalEntry = journalEntryOptional.get();
+        if(journalEntry.getDnevnikKorisnikId() != userId) {
+            throw new MethodNotAllowedException();
+        }
+
         journalEntryRepository.deleteById(id);
     }
